# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse:
    properties:
      error:
        type: string
      fields: {}
    type: object
  github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse:
    properties:
      data: {}
    type: object
...
  internal_handler.CreateCandidateRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    type: object
...
  internal_handler.CreateOfferRequest:
    properties:
      candidate_id:
        type: string
      created_by:
        type: string
      department_id:
        type: string
      designation_id:
        type: string
      level_id:
        type: string
      new_start_date:
        type: string
      offer_letter_url:
        type: string
      proposed_start_date:
        type: string
    required:
    - candidate_id
    - created_by
    - department_id
    - designation_id
    - level_id
    - proposed_start_date
    type: object
...
  internal_handler.UpdateCandidateRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    required:
    - first_name
    - last_name
    type: object
...
info:
  contact: {}
  description: REST API documentation for Workaround backend services.
  title: Workaround Backend API
  version: "1.0"
paths:
...
  /v1/candidates:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of candidates
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: List candidates
      tags:
      - Candidates
    post:
      consumes:
      - application/json
      description: Creates a new candidate
      parameters:
      - description: Candidate payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateCandidateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Create candidate
      tags:
      - Candidates
  /v1/candidates/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a candidate by id
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Delete candidate
      tags:
      - Candidates
    get:
      consumes:
      - application/json
      description: Retrieves a candidate by id
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Get candidate by id
      tags:
      - Candidates
    put:
      consumes:
      - application/json
      description: Updates a candidate by id
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      - description: Candidate payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateCandidateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Update candidate
      tags:
      - Candidates
  /v1/candidates/{id}/offers:
    get:
      consumes:
      - application/json
      description: Retrieves offers for a candidate
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Get offers by candidate
      tags:
      - Offers
...
  /v1/offers:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of offers
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: List offers
      tags:
      - Offers
    post:
      consumes:
      - application/json
      description: Creates a new candidate offer
      parameters:
      - description: Offer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateOfferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Create offer
      tags:
      - Offers
  /v1/offers/{id}/accept:
    put:
      consumes:
      - application/json
      description: Accepts a candidate offer by id
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Accept offer
      tags:
      - Offers
  /v1/offers/{id}/reject:
    put:
      consumes:
      - application/json
      description: Rejects a candidate offer by id
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_falasefemi2_workaround-backend_internal_response.ErrorResponse'
      summary: Reject offer
      tags:
      - Offers
...
schemes:
- http
- https
swagger: "2.0"